Update docker version

+++ build/scripts/install-app.sh
@@ -2,6 +2,11 @@
 
 APP_TMP_DIR="build/out/app-tmp"
 
+if [ -d "./app/vendor" ]; then
+  echo "The app is already installed."
+  exit 0
+fi
+
 docker run --rm -u "$(id -u)":"$(id -g)" -v "$(pwd)":/app composer create-project symfony/skeleton:"6.4.*" $APP_TMP_DIR
 
 docker run --rm -u "$(id -u)":"$(id -g)" -v "$(pwd)/$APP_TMP_DIR":/app composer:lts composer require --dev friendsofphp/php-cs-fixer
